+++ java-client-sdk/lib/src/main/java/industries/dot/brease/ContextID.java
@@ -49,7 +49,8 @@ public class ContextID {
         req.setURL(url);
         SerializedBody serializedRequestBody = industries.dot.brease.utils.Utils.serializeRequestBody(request, "addRuleInput", "json");
         req.setBody(serializedRequestBody);
-        
+
+        req.addHeader("Accept", "application/json");
         req.addHeader("user-agent", String.format("speakeasy-sdk/%s %s %s", this._language, this._sdkVersion, this._genVersion));
         
         HTTPClient client = industries.dot.brease.utils.Utils.configureSecurityClient(this._defaultClient, security);
@@ -90,7 +91,8 @@ public class ContextID {
         req.setURL(url);
         SerializedBody serializedRequestBody = industries.dot.brease.utils.Utils.serializeRequestBody(request, "evaluateRulesInput", "json");
         req.setBody(serializedRequestBody);
-        
+
+        req.addHeader("Accept", "application/json");
         req.addHeader("user-agent", String.format("speakeasy-sdk/%s %s %s", this._language, this._sdkVersion, this._genVersion));
         
         HTTPClient client = industries.dot.brease.utils.Utils.configureSecurityClient(this._defaultClient, security);
@@ -129,7 +131,8 @@ public class ContextID {
         HTTPRequest req = new HTTPRequest();
         req.setMethod("GET");
         req.setURL(url);
-        
+
+        req.addHeader("Accept", "application/json");
         req.addHeader("user-agent", String.format("speakeasy-sdk/%s %s %s", this._language, this._sdkVersion, this._genVersion));
         
         HTTPClient client = industries.dot.brease.utils.Utils.configureSecurityClient(this._defaultClient, security);
@@ -168,7 +171,8 @@ public class ContextID {
         HTTPRequest req = new HTTPRequest();
         req.setMethod("DELETE");
         req.setURL(url);
-        
+
+        req.addHeader("Accept", "*/*");
         req.addHeader("user-agent", String.format("speakeasy-sdk/%s %s %s", this._language, this._sdkVersion, this._genVersion));
         
         HTTPClient client = industries.dot.brease.utils.Utils.configureSecurityClient(this._defaultClient, security);
@@ -203,7 +207,8 @@ public class ContextID {
         req.setURL(url);
         SerializedBody serializedRequestBody = industries.dot.brease.utils.Utils.serializeRequestBody(request, "replaceRuleInput", "json");
         req.setBody(serializedRequestBody);
-        
+
+        req.addHeader("Accept", "application/json");
         req.addHeader("user-agent", String.format("speakeasy-sdk/%s %s %s", this._language, this._sdkVersion, this._genVersion));
         
         HTTPClient client = industries.dot.brease.utils.Utils.configureSecurityClient(this._defaultClient, security);