+++ python-client-sdk/src/sdk/contextid.py
@@ -32,6 +32,7 @@ class ContextID:
         req_content_type, data, form = utils.serialize_request_body(request, "add_rule_input", 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
+        headers['Accept'] = 'application/json'
         headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)
@@ -58,6 +59,7 @@ class ContextID:
         req_content_type, data, form = utils.serialize_request_body(request, "evaluate_rules_input", 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
+        headers['Accept'] = 'application/json'
         headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)
@@ -81,6 +83,7 @@ class ContextID:
         
         url = utils.generate_url(operations.GetAllRulesRequest, base_url, '/{contextID}/rules', request)
         headers = {}
+        headers['Accept'] = 'application/json'
         headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)
@@ -104,6 +107,7 @@ class ContextID:
         
         url = utils.generate_url(operations.RemoveRuleRequest, base_url, '/{contextID}/rules/{id}', request)
         headers = {}
+        headers['Accept'] = '*/*'
         headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)
@@ -126,6 +130,7 @@ class ContextID:
         req_content_type, data, form = utils.serialize_request_body(request, "replace_rule_input", 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
+        headers['Accept'] = 'application/json'
         headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)