+++ python-client-sdk/src/sdk/models/shared/condition.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 from ..shared import conditionbaseref as shared_conditionbaseref
-from ..shared import conditiontype_enum as shared_conditiontype_enum
+from ..shared import conditiontype as shared_conditiontype
 from dataclasses_json import Undefined, dataclass_json
 from sdk import utils
 from typing import Any, Optional
@@ -22,5 +22,5 @@ class Condition:
     
     base: Optional[Any] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('base'), 'exclude': lambda f: f is None }})
     parameter: Optional[Any] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('parameter'), 'exclude': lambda f: f is None }})
-    type: Optional[shared_conditiontype_enum.ConditionTypeEnum] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})
+    type: Optional[shared_conditiontype.ConditionType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})